Create RandomDicesRoll
That's the first "concept" and the beginning, here it's where the magic happens, in this part of the program, like the file name says, it's where the dices are "Rolled" and generates the randomic value of the "dice" (the number of the dice it's the MAX number you can get and the minimum, like a regular dice, it's one). Also, the program is able to handle multiples dices and sum them, if the users want, they can get the biggest or the lowest number, and for last but not least, they can add a modifier to the dice.